requirements.txt
streamlit
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-groq
python-dotenv
